@@ -16,9 +16,11 @@
 //
 ////////////////////////////////////////////////////////////////////////////
 
-const { expect } = require("chai");
+import { expect } from "chai";
 
-const PersonAndDogsSchema = require("./schemas/person-and-dogs");
+import { IPerson, PersonAndDogSchema } from "./schemas/person-and-dogs";
+
+const RealmAsAny = Realm as any;
 
 describe("Realm#constructor", () => {
     it("is a function", () => {
@@ -48,22 +50,26 @@ describe("Realm#constructor", () => {
         expect(fileExists).to.equal(true);
         // Expect that the path is in the default path
         const defaultPathDir = path.dirname(Realm.defaultPath);
-        expect(realm1.path).to.equal(path.resolve(defaultPathDir, "temporary-1.realm"));
+        expect(realm1.path).to.equal(
+            path.resolve(defaultPathDir, "temporary-1.realm")
+        );
         // Open another Realm as well
         const realm2 = new Realm("temporary-2.realm");
-        expect(realm2.path).to.equal(path.resolve(defaultPathDir, "temporary-2.realm"));
+        expect(realm2.path).to.equal(
+            path.resolve(defaultPathDir, "temporary-2.realm")
+        );
     });
 
     describe("called with invalid arguments", () => {
         it("throws when called with an empty string", () => {
             expect(() => {
-                new Realm("");
+                const r = new Realm("");
             }).to.throw(); // The actual message varies across environments
         });
 
         it("throws when called with two strings", () => {
             expect(() => {
-                new Realm("", "");
+                const r = new RealmAsAny("", "");
             }).to.throw("Invalid arguments when constructing 'Realm'");
         });
     });
@@ -111,15 +117,18 @@ describe("Realm#constructor", () => {
             expect(realm1.schema.length).to.equal(0);
             realm1.close();
             // Re-open with a different version
-            const realm2 = new Realm({ schema: PersonAndDogsSchema, schemaVersion: 2 });
+            const realm2 = new Realm({
+                schema: PersonAndDogSchema,
+                schemaVersion: 2
+            });
             expect(realm2.schemaVersion).to.equal(2);
             expect(realm2.schema.length).to.equal(2);
             // Add an object, using the schema
             realm2.write(() => {
-                realm2.create("Person", { name: "John Doe", age: 42 });
+                realm2.create<IPerson>("Person", { name: "John Doe", age: 42 });
             });
             // Expect an object
-            const persons = realm2.objects("Person");
+            const persons = realm2.objects<IPerson>("Person");
             expect(persons.length).to.equal(1);
             expect(persons[0].name).to.equal("John Doe");
             expect(persons[0].age).to.equal(42);
@@ -131,7 +140,7 @@ describe("Realm#constructor", () => {
             expect(realm.schemaVersion).to.equal(0);
             expect(() => {
                 // Re-open with a different version
-                new Realm({ schema: [], schemaVersion: 1 });
+                const r = new Realm({ schema: [], schemaVersion: 1 });
             }).to.throw("already opened with different schema version");
         });
     });
@@ -139,7 +148,7 @@ describe("Realm#constructor", () => {
     describe("re-opening without a schema", () => {
         it("has the same schema as before", () => {
             // Open the Realm with a schema
-            const realm = new Realm({ schema: PersonAndDogsSchema });
+            const realm = new Realm({ schema: PersonAndDogSchema });
             realm.close();
             // Re-open it without a schema
             const reopenedRealm = new Realm();
@@ -152,98 +161,131 @@ describe("Realm#constructor", () => {
     describe("schema validation", () => {
         it("fails when passed an object", () => {
             expect(() => {
-                new Realm({ schema: {} });
+                const r = new RealmAsAny({ schema: {} });
             }).throws("schema must be of type 'array', got");
         });
 
         it("fails when passed an array with non-objects", () => {
             expect(() => {
-                new Realm({ schema: [ "" ] });
-            }).throws("Failed to read ObjectSchema: JS value must be of type 'object', got");
+                const r = new RealmAsAny({ schema: [""] });
+            }).throws(
+                "Failed to read ObjectSchema: JS value must be of type 'object', got"
+            );
         });
 
         it("fails when passed an array with empty object", () => {
             expect(() => {
-                new Realm({ schema: [ {} ] });
-            }).throws("Failed to read ObjectSchema: name must be of type 'string', got ");
+                const r = new RealmAsAny({ schema: [{}] });
+            }).throws(
+                "Failed to read ObjectSchema: name must be of type 'string', got "
+            );
         });
 
         it("fails when passed an array with an object without 'properties'", () => {
             expect(() => {
-                new Realm({ schema: [ { name: "SomeObject" } ] });
-            }).throws("Failed to read ObjectSchema: properties must be of type 'object', got ");
+                const r = new RealmAsAny({ schema: [{ name: "SomeObject" }] });
+            }).throws(
+                "Failed to read ObjectSchema: properties must be of type 'object', got "
+            );
         });
 
         it("fails when passed an array with an object without 'name'", () => {
             expect(() => {
-                new Realm({ schema: [ { properties: {} } ] });
-            }).throws("Failed to read ObjectSchema: name must be of type 'string', got ");
+                const r = new RealmAsAny({ schema: [{ properties: {} }] });
+            }).throws(
+                "Failed to read ObjectSchema: name must be of type 'string', got "
+            );
         });
 
-        function expectInvalidProperty(property, message) {
+        function expectInvalidProperty(
+            badProperty: Realm.PropertyType | Realm.ObjectSchemaProperty,
+            message: string
+        ) {
             expect(() => {
-                new Realm({ schema: [
-                    { name: "InvalidObject", properties: {
-                        bad: property,
-                        nummeric: "int",
-                        another: "AnotherObject"
-                    } },
-                    { name: "AnotherObject", properties: {} }
-                ]});
+                const r = new Realm({
+                    schema: [
+                        {
+                            name: "InvalidObject",
+                            properties: {
+                                another: "AnotherObject",
+                                bad: badProperty,
+                                nummeric: "int"
+                            }
+                        },
+                        { name: "AnotherObject", properties: {} }
+                    ]
+                });
             }).throws(message);
         }
 
         it("fails when asking for a list of lists", () => {
             expectInvalidProperty(
-                { type:"list[]", objectType: "InvalidObject" },
+                { type: "list[]", objectType: "InvalidObject" },
                 "List property 'InvalidObject.bad' must have a non-list value type"
             );
         });
 
         it("fails when asking for an optional list", () => {
             expectInvalidProperty(
-                { type:"list?", objectType: "InvalidObject" },
+                { type: "list?", objectType: "InvalidObject" },
                 "List property 'InvalidObject.bad' cannot be optional"
             );
         });
 
         it("fails when asking for an empty type string", () => {
-            expectInvalidProperty("", "Property 'InvalidObject.bad' must have a non-empty type");
+            expectInvalidProperty(
+                "",
+                "Property 'InvalidObject.bad' must have a non-empty type"
+            );
         });
 
         it("fails when asking for linkingObjects to a non-existing property", () => {
-            expectInvalidProperty({
-                type: "linkingObjects",
-                objectType: "InvalidObject",
-                property: "nosuchproperty"
-            }, "Property 'InvalidObject.nosuchproperty' declared as origin of linking objects property 'InvalidObject.bad' does not exist");
+            expectInvalidProperty(
+                {
+                    objectType: "InvalidObject",
+                    property: "nosuchproperty",
+                    type: "linkingObjects"
+                },
+                "Property 'InvalidObject.nosuchproperty' declared as origin of linking objects property 'InvalidObject.bad' does not exist"
+            );
         });
 
         it("fails when asking for linkingObjects to a non-link property", () => {
-            expectInvalidProperty({
-                type: "linkingObjects",
-                objectType: "InvalidObject",
-                property: "nummeric"
-            }, "Property 'InvalidObject.nummeric' declared as origin of linking objects property 'InvalidObject.bad' is not a link");
+            expectInvalidProperty(
+                {
+                    objectType: "InvalidObject",
+                    property: "nummeric",
+                    type: "linkingObjects"
+                },
+                "Property 'InvalidObject.nummeric' declared as origin of linking objects property 'InvalidObject.bad' is not a link"
+            );
         });
 
         it("fails when asking for linkingObjects to a property linking elsewhere", () => {
-            expectInvalidProperty({
-                type: "linkingObjects",
-                objectType: "InvalidObject",
-                property: "another"
-            }, "Property 'InvalidObject.another' declared as origin of linking objects property 'InvalidObject.bad' links to type 'AnotherObject'");
+            expectInvalidProperty(
+                {
+                    objectType: "InvalidObject",
+                    property: "another",
+                    type: "linkingObjects"
+                },
+                "Property 'InvalidObject.another' declared as origin of linking objects property 'InvalidObject.bad' links to type 'AnotherObject'"
+            );
         });
 
         it("allows list of objects with objectType defined", () => {
-            new Realm({ schema: [
-                { name: "SomeObject", properties: {
-                    myObjects: {
-                        type: "object[]",
-                        objectType: "SomeObject"
+            const r = new Realm({
+                schema: [
+                    {
+                        name: "SomeObject",
+                        properties: {
+                            myObjects: {
+                                objectType: "SomeObject",
+                                type: "object[]"
+                            }
+                        }
                     }
-                } },
-            ]});
+                ]
+            });
         });
     });
 
@@ -253,9 +295,9 @@ describe("Realm#constructor", () => {
 // Testing static methods
 
 describe("#deleteFile", () => {
-    function expectDeletion(path) {
+    function expectDeletion(path?: string) {
         // Create the Realm with a schema
-        const realm = new Realm({ path, schema: PersonAndDogsSchema });
+        const realm = new Realm({ path, schema: PersonAndDogSchema });
         realm.close();
         // Delete the Realm
         Realm.deleteFile({ path });
@@ -263,6 +305,7 @@ describe("#deleteFile", () => {
         const reopenedRealm = new Realm({ path });
         expect(reopenedRealm.schema).to.deep.equal([]);
     }
+
     it("deletes the default Realm", () => {
         expectDeletion();
     });
@@ -270,7 +313,7 @@ describe("#deleteFile", () => {
     // TODO: Fix the issue on Android that prevents this from passing
     // @see https://github.com/realm/realm-js-private/issues/507
 
-    it.skipIf('android', "deletes a Realm with a space in its path", () => {
+    it.skipIf("android", "deletes a Realm with a space in its path", () => {
         expectDeletion("my realm.realm");
     });
 });
