@@ -90,6 +90,9 @@ function fetchWithTimeout(input, init) {
                 url: xhr.responseURL,
                 headers: xhr.responseHeaders
             };
+            if (!options.headers) {
+                options.headers = {'content-type': xhr.getResponseHeader('content-type')};
+            }
             const body = 'response' in xhr ? xhr.response : xhr.responseText;
             resolve(new Response(body, options));
         };
@@ -97,7 +100,7 @@ function fetchWithTimeout(input, init) {
         xhr.ontimeout = () => reject(new TypeError('Network request failed'));
         xhr.open(request.method, request.url, true);
         request.headers.forEach((value, name) => xhr.setRequestHeader(name, value));
-        xhr.send(typeof request._bodyInit === 'undefined' ? null : request._bodyInit);
+        xhr.send(typeof request._bodyInit === 'undefined' ? init.body : request._bodyInit);
     });
 }
 
